FrontEnd2/requirements.txt

pygame==2.0
python-socketio[client]
